refactor(architecture): apply hexagonal architecture to project structure Refs #4
