+++ frontend/app/index.tsx
@@ -1,15 +1,2 @@
-import { Text, View } from "react-native";
-
-export default function Index() {
-  return (
-    <View
-      style={{
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-      }}
-    >
-      <Text>Hello world!</Text>
-    </View>
-  );
-}
+import landing from "@/interfaces/screens/landing";
+export default landing;
