+++ frontend/src/interfaces/layouts/mainLayout.tsx
@@ -1,0 +1,5 @@
+import { Stack } from "expo-router";
+
+export default function RootLayout() {
+  return <Stack />;
+}
